Activates BitLocker
This script checks to see if bitlocker can be enabled, intializes the tpm if necessary, and enables bitlocker with logging. It will require some tweaking or a companion script to store the recovery key in AD, and I am working on that. This script is best deployed by an RMM, but can also be run at the end of a task sequence, or by psremoting/pssession which will require the execution policy to be changed to unsigned or bypass.